refactor(react-app): 移除 admin 相关功能(后续归 admin 子包管理)
