+++ packages/react-app/src/App.tsx
@@ -8,11 +8,9 @@ import { difyChatRuntimeConfig } from '@/config/global'
 
 import './App.css'
 import LayoutIndex from './layout'
-import AppManagementPage from './pages/admin/app-management'
-import SystemConfigPage from './pages/admin/system-config'
-import AppListPage from './pages/client/apps'
-import AuthPage from './pages/client/auth'
-import ChatPage from './pages/client/chat'
+import AppListPage from './pages/apps'
+import AuthPage from './pages/auth'
+import ChatPage from './pages/chat'
 
 // 初始化响应式配置
 initResponsiveConfig()
@@ -21,12 +19,10 @@ initResponsiveConfig()
 difyChatRuntimeConfig.init('multiApp')
 
 const routes: IRoute[] = [
-	{ path: '/admin/app-management', component: () => <AppManagementPage /> },
-	{ path: '/admin/system-config', component: () => <SystemConfigPage /> },
-	{ path: '/client/auth', component: () => <AuthPage /> },
-	{ path: '/client/chat', component: () => <ChatPage /> },
-	{ path: '/client/app/:appId', component: () => <ChatPage /> },
-	{ path: '/client/apps', component: () => <AppListPage /> },
+	{ path: '/auth', component: () => <AuthPage /> },
+	{ path: '/chat', component: () => <ChatPage /> },
+	{ path: '/app/:appId', component: () => <ChatPage /> },
+	{ path: '/apps', component: () => <AppListPage /> },
 ]
 
 /**
